requirements.txt
tornado
RPLCD
RPi.GPIO
pytest
pytest-cov